- add sign up page styles - split form to component
