+++ src/components/ValidationError.js
@@ -3,14 +3,14 @@ import PropTypes from 'prop-types';
 
 const ValidationError = ({ inputField, isValid, error }) => {
   const errors = error.map((err) => (
-    <p key={err} className="text-red-600">
+    <p key={err} className="text-red-600 text-xs">
       *&nbsp;
       {inputField}
       &nbsp;
       {err}
     </p>
   ));
-  return <div className="h-6">{isValid ? null : errors}</div>;
+  return <div className="px-3">{isValid ? null : errors}</div>;
 };
 
 ValidationError.defaultProps = {
