+++ src/components/auth/SignUp.js
@@ -1,111 +1,19 @@
-import React, { useState } from 'react';
-import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
-import PropTypes from 'prop-types';
-import ReactRouterPropTypes from 'react-router-prop-types';
-import ValidationError from '../ValidationError';
-import { signUpWithUserData } from '../../actions/asyncActions';
+import React from 'react';
+import SignUpForm from './SignUpForm';
 
-const SignUp = ({ signUpWithUserData, validationErrors, history }) => {
-  const [inputFields, setInputFields] = useState({
-    email: '',
-    first_name: '',
-    last_name: '',
-    password: '',
-    password_confirmation: ''
-  });
+const SignUp = () => (
+  <div className="flex items-center w-full justify-center h-screen bg-teal-100 font-montserrat">
+    <div className="w-4/5 max-w-md">
+      <div className="text-center my-8">
+        <h2 className="text-teal-500 font-bold text-3xl leading-loose">
+          <span className="text-gray-700">I</span>
+          Doc
+        </h2>
+        <h3 className="font-bold uppercase">Sign up to continue</h3>
+      </div>
+      <SignUpForm />
+    </div>
+  </div>
+);
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    signUpWithUserData(inputFields, history);
-  };
-
-  return (
-    <form className="flex flex-col" onSubmit={handleSubmit} noValidate>
-      <input
-        type="email"
-        placeholder="Email"
-        value={inputFields.email}
-        // prettier-ignore
-        onChange={(e) => setInputFields({ ...inputFields, email: e.target.value })}
-        className="p-1 border-2"
-      />
-      <ValidationError
-        inputField="Email"
-        isValid={validationErrors.isValid}
-        error={validationErrors.errors.email}
-      />
-      <input
-        type="text"
-        placeholder="First Name"
-        value={inputFields.first_name}
-        // prettier-ignore
-        onChange={(e) => setInputFields({ ...inputFields, first_name: e.target.value })}
-        className="p-1 border-2"
-      />
-      <ValidationError
-        inputField="First Name"
-        isValid={validationErrors.isValid}
-        error={validationErrors.errors.first_name}
-      />
-      <input
-        type="text"
-        placeholder="Last Name"
-        value={inputFields.last_name}
-        // prettier-ignore
-        onChange={(e) => setInputFields({ ...inputFields, last_name: e.target.value })}
-        className="p-1 border-2"
-      />
-      <ValidationError
-        inputField="Last Name"
-        isValid={validationErrors.isValid}
-        error={validationErrors.errors.last_name}
-      />
-      <input
-        type="password"
-        autoComplete="password"
-        placeholder="Password"
-        value={inputFields.password}
-        // prettier-ignore
-        onChange={(e) => setInputFields({ ...inputFields, password: e.target.value })}
-        className="p-1 border-2"
-      />
-      <ValidationError
-        inputField="Password"
-        isValid={validationErrors.isValid}
-        error={validationErrors.errors.password}
-      />
-      <input
-        type="password"
-        autoComplete="password"
-        placeholder="Password Confirmation"
-        value={inputFields.password_confirmation}
-        // prettier-ignore
-        onChange={(e) => setInputFields({ ...inputFields, password_confirmation: e.target.value })}
-        className="p-1 border-2"
-      />
-      <ValidationError
-        inputField="Password Confirmation"
-        isValid={validationErrors.isValid}
-        error={validationErrors.errors.password_confirmation}
-      />
-      <button type="submit">Create Account</button>
-    </form>
-  );
-};
-
-SignUp.propTypes = {
-  history: ReactRouterPropTypes.history.isRequired,
-  signUpWithUserData: PropTypes.func.isRequired,
-  validationErrors: PropTypes.shape({
-    isValid: PropTypes.bool.isRequired,
-    errors: PropTypes.objectOf(PropTypes.arrayOf(PropTypes.string))
-  }).isRequired
-};
-
-const mapStateToProps = (state) => ({
-  notification: state.notification,
-  validationErrors: state.validationErrors
-});
-
-export default connect(mapStateToProps, { signUpWithUserData })(withRouter(SignUp));
+export default SignUp;
